feat: implementado o tratamento de exceções na classe SerieService

+++ src/main/java/com/trybe/acc/java/minhasseries/service/SerieService.java
@@ -1,5 +1,7 @@
 package com.trybe.acc.java.minhasseries.service;
 
+import com.trybe.acc.java.minhasseries.exception.SerieExistenteException;
+import com.trybe.acc.java.minhasseries.exception.SerieNaoEncontradaException;
 import com.trybe.acc.java.minhasseries.model.Episodio;
 import com.trybe.acc.java.minhasseries.model.Serie;
 import com.trybe.acc.java.minhasseries.repository.SerieRepository;
@@ -18,7 +20,15 @@ public class SerieService {
   @Autowired
   private SerieRepository serieRepository;
 
+  /**
+   * Método que cria uma série.
+   * 
+   */
   public Serie createSerie(Serie serie) {
+    if (serieRepository.existsByNome(serie.getNome())) {
+      throw new SerieExistenteException();
+    }
+  
     return serieRepository.save(serie);
   }
 
@@ -26,7 +36,15 @@ public class SerieService {
     return serieRepository.findAll();
   }
 
+  /**
+   * Método que deleta uma série.
+   * 
+   */
   public String deleteSerie(Integer serieId) {
+    if (!serieRepository.existsById(serieId)) {
+      throw new SerieNaoEncontradaException();
+    }
+
     serieRepository.deleteById(serieId);
     return "Serie deletada com sucesso";
   }
@@ -36,10 +54,16 @@ public class SerieService {
    * 
    */
   public Serie addEpisode(Integer serieId, Episodio episodio) {
-    Serie serie = serieRepository.findById(serieId).get();
-    episodio.setSerie(serie);
-    serie.adicionarEpisodio(episodio);
-    return serieRepository.save(serie);
+    try {
+      Serie serie = serieRepository.findById(serieId).get();
+
+      episodio.setSerie(serie);
+      serie.adicionarEpisodio(episodio);
+      return serieRepository.save(serie);
+    } catch (RuntimeException e) {
+      throw new SerieNaoEncontradaException();
+    }
+
   }
 
   public List<Episodio> getEpisodes(Integer serieId) {
